--- .nyc_output/processinfo/417affc5-d6c6-43f1-aa8d-865034648968.json
@@ -1,1 +1,0 @@
-{"uuid":"417affc5-d6c6-43f1-aa8d-865034648968","parent":null,"pid":8380,"argv":["C:\\Program Files\\nodejs\\node.exe","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\node_modules\\mocha\\bin\\mocha","./src/test/index.js","--timeout","20000","--exit"],"execArgv":[],"cwd":"E:\\de\\teamwork-fullstack-app\\teamwork-backend","time":1574049236973,"ppid":8524,"root":"5eb32a49-e458-4b5b-9f92-1c317e858289","coverageFilename":"E:\\de\\teamwork-fullstack-app\\teamwork-backend\\.nyc_output\\417affc5-d6c6-43f1-aa8d-865034648968.json","files":["E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\index.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\migrations\\index.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\db\\index.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\config\\config.js"]}
