+++ .nyc_output/processinfo/96cda833-6742-4067-8b1e-bb3fa6075a04.json
@@ -1,0 +1,1 @@
+{"uuid":"96cda833-6742-4067-8b1e-bb3fa6075a04","parent":null,"pid":8756,"argv":["C:\\Program Files\\nodejs\\node.exe","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\node_modules\\mocha\\bin\\mocha","./src/test/","--timeout","40000","--exit"],"execArgv":[],"cwd":"E:\\de\\teamwork-fullstack-app\\teamwork-backend","time":1574072056811,"ppid":8744,"root":"a4bc652a-8fe6-4856-a2c5-c651c29dee1d","coverageFilename":"E:\\de\\teamwork-fullstack-app\\teamwork-backend\\.nyc_output\\96cda833-6742-4067-8b1e-bb3fa6075a04.json","files":["E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-a--create-article.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\app.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\routes\\auth.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\middleware\\input-validator.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\middleware\\auth.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\utilities\\responseUtility.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\controllers\\auth.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\db\\index.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\config\\config.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\routes\\articles.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\controllers\\articles.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-aa--signin.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-ab--create-user.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-b--get-articles-by-tag.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-c--get-article-by-id.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-d--edit-article.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-e--flag-article.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\utilities\\testQueryUtility.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-f--delete-article.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-g--delete-flagged-article.js","E:\\de\\teamwork-fullstack-app\\teamwork-backend\\src\\test\\test-h--comment-artcle.js"]}
