+++ package.json
@@ -6,8 +6,7 @@
   "scripts": {
     "build": "rm -rf ./build && babel -d build/src ./src -s",
     "dev": "nodemon --exec babel-node src/server.js",
-    "datadump": "nyc --require @babel/register  mocha ./src/test/datadump.js --timeout 20000 --exit",
-    "test": "nyc --require @babel/register  mocha ./src/test/api/articles/ --timeout 40000 --exit",
+    "test": "nyc --require @babel/register  mocha ./src/test/ --timeout 40000 --exit",
     "generate-lcov": "nyc report --reporter=text-lcov > lcov.info",
     "coveralls-coverage": "coveralls < lcov.info",
     "coverage": "npm run generate-lcov && npm run coveralls-coverage"
